Break up scalems.radical modules.
Move everything but RPDispatchingExecutor from
scalems.radical.runtime to new modules:
* configuration
* exceptions
* manager
* session
* task

Update `scalems.radical.configuration()` semantics:
* Reduce exposure of `_configuration` ContextVar.
  - The function no longer sets any state. It is just a factory.
  - `current_configuration()` can query the scalems.radical.runtime
    module state.
  - Module state is manipulated only by
    `RPDispatchingExecutor.runtime_configuration()` context manager.
* Use `configuration()` to invoke the parser.
* Use `configuration(*args, **kwargs)` as a lower level utility.

Separate the datastore from the configuration.

Move `get_scheduler()` to public scalems.radical.raptor module function.
